GLT-4423 Concentration Feild Size increment

+++ miso-web/src/main/webapp/scripts/bulk_library.js
@@ -960,7 +960,7 @@ BulkTarget.library = (function ($) {
           title: "Spike-In Volume",
           type: "decimal",
           data: "spikeInVolume",
-          precision: 14,
+          precision: 16,
           scale: 10,
         }
       );
